Abstract crypto logic to separate class
